refactor: centralized form logic and content moving removing specific form duplication
